requirements.txt
requests >= 2.28.1
pytest >= 7.2.0
voluptuous >= 0.13.1
pytest-voluptuous >= 1.2.0
python-dotenv >= 0.21.0
allure-pytest >= 2.12.0
curlify >= 2.2.1